HACK: Try reverse to work around bug #15407
The reported bug[1] reported a problem when using "**" as branch specifier with
github and provide some tests which show that the real problem is on the git
plugin.

Looking the branches that were built, I thought this patch may simply work
around that. This patch is NOT, by any means, a fix for the real problem (it
introduces other problems). But is a "proof of concept" to:

1) Be more sure that the bug is on the git plugin (although tests were pretty
conclusive too)

2) Have an ugly work around to better understand the problem

But a real fix for the problem is needed.

[1]: https://issues.jenkins-ci.org/browse/JENKINS-15407?page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel